Add functionality that lets user pick grid size
A button (New Grid) can be pressed by the user and will prompt a number.
This number is used as the new grid dimensions, ex: user inputs 20 and
the grid size will be 20x20.

Also separating logic into functions; the initial grid simply
calls on the makeGrid instead of having the for loop exposed raw, and
the button clears the div children. Now it only needs to match the CSS.