Change grid sizing to be static
Previously the grid's boundaries would change based on grid size.
However, this was due to my lack of understanding the Odin Project
assignment. The grid has been changed to reflect this: regardless of the
grid's column and row sizes, the total dimensions of the grid should
stay the same.

As a result, I have arbitrarily set the grid's height and width to 720px
and thus the child divs within should change size to fit this container.
For example, a 16x16 grid will have larger boxes while a 100x100 grid
should have tiny boxes.