Organize divs into an actual 16 x 16 grid
Rather than hardcoding dimensions, the container's height and width are
programmatically determined based on the size of the children divs. This
way it is possible to change the children div's values in CSS and not
break the alignment of the grid. This will also be useful later when we
add a feature to let the user pick their own grid size.