Rework grid layout to be based on container size
Previously the container height/width was calculated in JS by obtaining
the size of the individual units. This wound up being undesirable
as this means the container gets bigger and smaller based on number of
units, taking up more page space.

Instead it was opted to keep the container at a fixed size, and thus
the unit size will be calculated using that. As a result, dependent on
the desired grid size, the individual units will get a different height
and width programmactically in JS.

Originally I tried to devise a way to make the units change size through
flexing and flex-wrapping, but it wound up being a major headache.